% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\alias{balance_ATE}
\alias{balance_ATT}
\alias{balance_OWATE}
\title{Covariate Balancing Weights via Generalized Projections of Bregman Distances}
\usage{
balance_ATE(
  X,
  Y,
  Z,
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)

balance_ATT(
  X,
  Y,
  Z,
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)

balance_OWATE(
  X,
  Y,
  Z,
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)
}
\arguments{
\item{X}{the balance functions to be contrained.}

\item{Y}{the observed responses.}

\item{Z}{the binary treatment assignment.}

\item{base_weights}{a vector of optional base weights with length equal to the
number of rows in \code{X}.}

\item{coefs_init}{the optional initial values for the dual variables. Default is a vector of zeros with length 
equal to number of columns in \code{X}.}

\item{optim_ctrl}{a list of arguments that will be passed to \code{optim()}.}

\item{...}{additional arguments.}
}
\description{
The \code{balance_ATE()}, \code{balance_ATT()}, and \code{balance_OWATE()} functions solve a convex program with linear equality constraints determined by the data, the
estimand (ATE, ATT, or OWATE), and the sampling weights (\code{base_weights}).
}
\references{
Josey KP, Juarez-Colunga E, Yang F, Ghosh D (2019). "A Framework for Covariate Balance using Bregman
Distances." arXiv:1903.00390 [stat].
}
