% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{bootit}
\alias{bootit}
\title{Bootstrap Indices}
\usage{
bootit(Z)
}
\arguments{
\item{Z}{treatment assignment vector.}
}
\description{
Produces a vector of indices which are utilized in the bootstrap
variance estimation functioniality of \code{cestimate()}.
}
