% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Constrained Convex Optimization of Bregman Distances}
\usage{
calibrate(
  constraint,
  target,
  distance = c("entropy", "binary", "shifted"),
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)
}
\arguments{
\item{constraint}{a matrix that forms the basis of a linear subspace which define the equality constraints of the
convex program.}

\item{target}{the target margins of the linear equality constraints. This vector
should have a length equal to the number of columns in \code{constraint}.}

\item{distance}{the Bregman distance to be optimized. Can either be "entropy" for the relative entropy,
"binary" for the binary relative entropy, or "shifted" for the shifted relative entropy.}

\item{base_weights}{a vector of optional base weights with length equal to the
number of rows in \code{constraint}.}

\item{coefs_init}{the optional initialization values for the dual variables. Default is a vector of zeros with length 
equal to number of columns in \code{X}.}

\item{optim_ctrl}{a list of arguments that will be passed to \code{optim()}.}

\item{...}{additional arguments.}
}
\description{
The \code{calibrate()} function solves a convex program with linear equality constraints determined by the 
constraint matrix \code{constraint}, the estimand (\code{estimand}), and the sampling weights (\code{base_weights}).
The function \code{calibrate()} provides a more direct means to solving the convex optimization program. However, 
the constraint matrix and target margins must be determined by the user.
}
\references{
Censor Y, Zenios SA (1998). Parallel Optimization: Theory, Algorithms, and Applications. 1st ed. New York:
Oxford University Press.
}
