% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{cbal_est}
\alias{cbal_est}
\title{Function for Estimating Causal Effects with \code{cbal}.}
\usage{
cbal_est(obj, Y, ...)
}
\arguments{
\item{obj}{object of class "cbalance" or "cboost".}

\item{Y}{numeric outcome vector.}
}
\description{
This function returns the Horvitz-Thompson estimates and the sandwich variance
estimate from an object of type \code{cbalance}.
}
