% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{esteq}
\alias{esteq}
\alias{esteq_ATE}
\alias{esteq_HTE}
\alias{esteq_transport}
\alias{esteq_fusion}
\title{Horvitz-Thompson Estimating Equations}
\usage{
esteq_ATE(X, Y, Z, p, base_weights, tau)

esteq_HTE(X, Y, Z, p, base_weights, tau)

esteq_transport(S, X, Y, Z, p, base_weights, theta, tau)

esteq_fusion(S, X, Y, Z, p, base_weights, theta, tau)
}
\arguments{
\item{X}{the balance functions to be contrained.}

\item{Y}{the observed responses.}

\item{Z}{the binary treatment assignment.}

\item{p}{the estimated balancing weights.}

\item{base_weights}{a vector of base weights with length equal to the 
number of rows in \code{X}.}

\item{tau}{causal effect estimate.}

\item{S}{the binary vector of sample indicators.}

\item{theta}{target sample means of the balance functions.}
}
\description{
This function is required for the sandwich variance estimator. It provides the
estimating equations for the "meat".
}
