% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{fusion}
\alias{fusion}
\alias{fusion_ATE}
\title{Balancing Weights for Data-Fusion}
\usage{
fusion_ATE(
  S,
  X,
  Y,
  Z,
  base_weights = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)
}
\arguments{
\item{S}{the binary vector of sample indicators.}

\item{X}{the balance functions to be contrained.}

\item{Y}{the observed responses.}

\item{Z}{the binary treatment assignment.}

\item{base_weights}{a vector of optional base weights with length equal to the
number of rows in \code{X}.}

\item{optim_ctrl}{a list of arguments that will be passed to \code{optim()}.}

\item{...}{additional arguments.}
}
\description{
The \code{fusion_ATE()} function finds balancing weights for combining datasets to estimate the target population average treatment effect.
This function requires complete individual-level data from both samples whereas the transport function only requires complete data from the
study sample and covariate data from the target sample.
}
\references{
Josey KP, Yang F, Ghosh D, Raghavan S (2020). "A Calibration Approach to Transportability and 
Data-Fusion with Observational Data." arXiv:2008.06615 [stat].
}
