% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{ks_data}
\alias{ks_data}
\title{Generate Kang & Schafer (2007) Scenarios}
\usage{
ks_data(tau, n, sig2, rho, y_scen = c("a", "b"), z_scen = c("a", "b"))
}
\arguments{
\item{tau}{the homogeneous causal effect to be estimated.}

\item{n}{the total sample size.}

\item{sig2}{the marginal error variance of the potential outcomes. We assume homoscedasticity.}

\item{rho}{the correlation coefficient between the potential outcomes.}

\item{y_scen}{the outcome scenario. Can be either \code{c("a", "b")}.}

\item{z_scen}{the treatment assignment scenario. Can be either \code{c("a", "b")}.}
}
\description{
The \code{ks_data()} function generates a dataset according to the experimental scenarios that appear in
Kang and Schafer (2007). These scenarios are useful for evauluating different covariate 
balancing methods under treatment and/or outcome model misspecification.
}
\references{
Kang, J. D., & Schafer, J. L. (2007). Demystifying double robustness: A
comparison of alternative strategies for estimating a population mean from
incomplete data. Statistical Science, 22(4), 523-539.
}
