% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagrange.R
\name{lagrange}
\alias{lagrange}
\alias{lagrange_ent}
\alias{lagrange_bent}
\alias{lagrange_sent}
\title{Langrangian Functions for Bregman Distances with Linear Equality Constraints}
\usage{
lagrange_ent(coefs, constr_mat, target_margins, base_weights)

lagrange_bent(coefs, constr_mat, target_margins, base_weights)

lagrange_sent(coefs, constr_mat, target_margins, base_weights)
}
\arguments{
\item{coefs}{vector of Lagrange multipliers.}

\item{constr_mat}{a matrix that determines the basis of a linear subspace where the equality constraints of the
optimization lie.}

\item{target_margins}{the target margins of the linear equality constraints. This vector 
should have a length equal to the number of columns in \code{constr_mat}.}

\item{base_weights}{a vector of optional sampling weights with length equal to the 
number of rows in \code{constr_mat}.}
}
\description{
Use \code{lagrange_ent} for the unnormalized relative entropy.
}
\details{
Use \code{lagrange_bent} for the binary relative entropyl

Use \code{lagrange_sent} for the shifted unnormalized relative entropyl
}
