% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{transport}
\alias{transport}
\alias{fusion}
\title{Balancing Weights for Transportability and Data-Fusion}
\usage{
transport(
  S,
  X,
  Z1,
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)

fusion(
  S,
  X,
  Z,
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)
}
\arguments{
\item{S}{the binary vector of sample indicators.}

\item{X}{the balance functions to be contrained.}

\item{Z1}{the binary treatment assignment for S = 1.}

\item{base_weights}{a vector of optional base weights with length equal to the
number of rows in \code{X}.}

\item{coefs_init}{the optional initialization values for the dual variables. Default is a vector of zeros with length 
equal to number of columns in \code{X}.}

\item{optim_ctrl}{a list of arguments that will be passed to \code{optim()}.}

\item{...}{additional arguments.}

\item{Z}{the binary treatment assignment.}
}
\description{
The \code{transport()} and \code{fusion()} functions find balancing weights for transporting estimates. The transport 
requires the complete individual-level data (response, treatment assignment, and covariates) for all units in the observed
sample but only the individual-level covariate data from the target sample. The fusion function requires complete individual-level
data from both samples.
}
\references{
Josey KP, Berkowitz SA, Ghosh D, Raghavan S (2020). "Transporting Experimental Results with Entropy
Balancing." arXiv:2002.07899 [stat].
}
