% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Covariate Balancing Weights via Generalized Projections of Bregman Distances}
\usage{
balance(
  X,
  Z,
  estimand = c("ATE", "ATT", "OWATE"),
  base_weights = NULL,
  coefs_init = NULL,
  optim_ctrl = list(maxit = 500, reltol = 1e-10),
  ...
)
}
\arguments{
\item{X}{the balance functions to be contrained.}

\item{Z}{the binary treatment assignment.}

\item{estimand}{the causal estimand to be estimated. Must be either "ATE" for the average treatment effect,
"ATT" for the average treatment effect, or "OWATE" for the optimally weighted average treatment effect.
In \code{balance()}, the estimand also determines the distance function supplied to \code{cfit()}.}

\item{base_weights}{a vector of optional base weights with length equal to the
number of rows in \code{X}.}

\item{coefs_init}{the optional initial values for the dual variables. Default is a vector of zeros with length 
equal to number of columns in \code{X}.}

\item{optim_ctrl}{a list of arguments that will be passed to \code{optim()}.}

\item{...}{additional arguments.}
}
\description{
The \code{balance()} function solves a convex program with linear equality constraints determined by the data, the
estimand (\code{estimand}), and the sampling weights (\code{base_weights}).
}
\references{
Josey KP, Juarez-Colunga E, Yang F, Ghosh D (2020). "A Framework for Covariate Balance using Bregman
Distances." Scandinavian Journal of Statistics, pp. 1-27. doi:10.1111/sjos.12457.
}
