% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{bootit}
\alias{bootit}
\title{Bootstrap Indices}
\usage{
bootit(Z, boot_frac)
}
\arguments{
\item{Z}{treatment assignment vector.}

\item{boot_frac}{The minimum proportion of entries from each treatment group represented
in the bootstrap resample.}
}
\description{
Produces a vector of indices which are utilized in the bootstrap
variance estimation functioniality of \code{cestimate()}.
}
