% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbalance.R
\name{cbalance.formula}
\alias{cbalance.formula}
\alias{cbalance.default}
\alias{cbalance}
\title{Covariate Balancing Weights via Generalized Projections of Bregman Distances}
\usage{
\method{cbalance}{formula}(formula, data, distance = c("entropy",
  "binary", "shifted"), estimand = c("ATE", "ATT", "ATC"),
  base_weights = NULL, coefs_init = NULL, control = list(maxit = 500,
  reltol = 1e-10), ...)

\method{cbalance}{default}(constr_mat, target_margins,
  base_weights = NULL, coefs_init = NULL, distance = c("entropy",
  "binary", "shifted"), control = list(maxit = 500, reltol = 1e-10), ...)

cbalance(...)
}
\arguments{
\item{formula}{an object of class formula: a symbolic description of the model to be fitted.}

\item{data}{a \code{data.frame}, list or environment containing the variables in the model.}

\item{distance}{the Bregman distance to be optimized. Can either be "entropy" for the relative entropy, 
"binary" for the binary relative entropy, or "shifted" for the shifted relative entropy.}

\item{estimand}{the assumed causal effect estimand. Can either be "ATE" for the average treatment effect,
"ATT" for the average treatment effect of the treated, or "ATC" for the average treatment effect of the controls.}

\item{base_weights}{a vector of optional sampling weights with length equal to the 
number of rows in \code{constr_mat} or \code{data}.}

\item{coefs_init}{initial values for the Lagrangian multipliers.}

\item{control}{a list of arguments that will be passed to \code{optim}.}

\item{...}{additional arguments.}

\item{constr_mat}{a matrix that forms the basis of a linear subspace where the equality constraints of the
optimization exist.}

\item{target_margins}{the target margins of the linear equality constraints. This vector 
should have a length equal to the number of columns in \code{constr_mat}.}
}
\description{
The \code{cbalance} function solves a convex program with linear equality constraints determined by the 
constraint matrix (\code{constr_mat}), the target margins (\code{target_margins}), and a Bregman distance (\code{distance}).
If \code{cbalance.formula} is used, then the constraint matrix and target margins are determined by either the ATE, ATT, or ATC.
}
\references{
Censor Y, Zenios SA (1998). Parallel Optimization: Theory, Algorithms, and Applications. 1st ed. New York:
Oxford University Press.
}
