% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbalance.R
\name{cbalance}
\alias{cbalance}
\alias{cbalance.fit}
\title{Covariate Balancing Weights via Generalized Projections of Bregman Distances}
\usage{
cbalance(formula, data, estimand = c("ATE", "ATT", "ATC"),
  distance = c("entropy", "binary", "shifted"), base_weights = NULL,
  coefs_init = NULL, degree = 1, control = list(maxit = 500, reltol =
  1e-10), ...)

cbalance.fit(constr_mat, target_margins, distance = c("entropy",
  "binary", "shifted"), base_weights = NULL, coefs_init = NULL,
  control = list(maxit = 500, reltol = 1e-10), ...)
}
\arguments{
\item{formula}{an object of class formula: a symbolic description of the model to be fitted.}

\item{data}{a \code{data.frame}, list, or environment containing the variables in the model.}

\item{estimand}{the assumed causal effect estimand. Can either be "ATE" for the average treatment effect,
"ATT" for the average treatment effect of the treated, or "ATC" for the average treatment effect of the controls.}

\item{distance}{the Bregman distance to be optimized. Can either be "entropy" for the relative entropy,
"binary" for the binary relative entropy, or "shifted" for the shifted relative entropy.}

\item{base_weights}{a vector of optional sampling weights with length equal to the
number of rows in \code{data} or \code{constr_mat}.}

\item{coefs_init}{initial values for the Lagrangian multipliers.}

\item{control}{a list of arguments that will be passed to \code{optim}.}

\item{...}{additional arguments.}

\item{constr_mat}{a matrix that forms the basis of a linear subspace which defines the equality 
constraints of the convex program.}

\item{target_margins}{the target margins of the linear equality constraints. This vector
should have a length equal to the number of columns in \code{constr_mat}.}
}
\description{
The \code{cbalance()} function solves a convex program with linear equality constraints determined by the
estimand (\code{estimand}), the criterion function (\code{distance}), and the sampling weights (\code{base_weights}).
The function \code{cbalance.fit()} provides a more direct means to solving the convex program. However, 
the constraint matrix and target margins must be determined by the user.
}
\references{
Censor Y, Zenios SA (1998). Parallel Optimization: Theory, Algorithms, and Applications. 1st ed. New York:
Oxford University Press.
}
