% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Function for Estimating Causal Effects using \code{balance}, \code{transport}, and \code{fusion} Objects}
\usage{
estimate(obj, Y, ...)
}
\arguments{
\item{obj}{object of class "balance", "transport", or "fusion".}

\item{Y}{the observed responses.}

\item{...}{additional arguments.}

\item{Y1}{the observed responses for S = 1.}
}
\description{
These functions return the causal effect estimates and the sandwich variance
estimates from an object of type "balance", "transport", or "fusion".
}
\references{
Josey KP, Juarez-Colunga E, Yang F, Ghosh D (2020b). "A Framework for Covariate Balance using Bregman
Distances." Scandinavian Journal of Statistics, pp. 1-27. doi:10.1111/sjos.12457.

Josey KP, Berkowitz SA, Ghosh D, Raghavan S (2020a). "Transporting Experimental Results with Entropy
Balancing." arXiv:2002.07899 [stat].
}
