% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{estimate}
\alias{estimate}
\alias{est_eq}
\title{Horvitz-Thompson Estimating Equations}
\usage{
est_eq(X, Y, Z, weights, tau)
}
\arguments{
\item{X}{design matrix.}

\item{Y}{outcome vector.}

\item{Z}{treatment assignment vector.}

\item{weights}{estimated balancing weights.}

\item{tau}{causal effect estimate.}
}
\description{
This function is required for the sandwich variance estimator. It provides the
estimating equations for the "meat".
}
